7_RAG-2/pyproject.toml
[tool.poetry]
name = "project1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.7"
python-dotenv = "^1.0.1"
langchain = "^0.2.5"
langchain-community = "^0.2.5"
langchain-core = "^0.2.9"
langchain-openai = "^0.1.9"
langchain-chroma = "^0.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
